+++ Cargo.toml
@@ -18,6 +18,7 @@ chrono = { version = "0.4.19", features = ["serde"] }
 serde_json = "1.0.64"
 semver = "1.0.24"
 uuid = { version = "1.13.2", features = ["serde", "v7"] }
+url = "2.5"
 
 [dev-dependencies]
 dotenv = "0.15.0"
